+++ packages/gamut-labs/src/AppHeaderMobile/AppHeaderMainMenuMobile/index.tsx
@@ -1,4 +1,10 @@
-import { ContentContainer, FillButton, TextButton } from '@codecademy/gamut';
+import {
+  Box,
+  ContentContainer,
+  FillButton,
+  FlexBox,
+  TextButton,
+} from '@codecademy/gamut';
 import React, { useState } from 'react';
 
 import { AppHeaderLink } from '../../AppHeader/AppHeaderElements/AppHeaderLink';
@@ -8,6 +14,7 @@ import {
   AppHeaderDropdownItem,
   AppHeaderItem,
 } from '../../AppHeader/AppHeaderElements/types';
+import { login, signUp } from '../../GlobalHeader/GlobalHeaderItems';
 import { AppHeaderSubMenuMobile } from '../AppHeaderSubMenuMobile';
 import { AppHeaderSubMenuTarget } from '../AppHeaderSubMenuTarget';
 import { MobileSearchBar } from './MobileSearchBar';
@@ -16,7 +23,8 @@ export type AppHeaderMainMenuMobileProps = {
   action: AppHeaderClickHandler;
   items: AppHeaderItem[];
   onSearch: (query: string) => void;
-  getItemType: (type: string) => void;
+  getItemType: (type: string | undefined) => void;
+  isAnon: boolean;
 };
 
 export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = ({
@@ -24,6 +32,7 @@ export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = (
   items,
   onSearch,
   getItemType,
+  isAnon,
 }) => {
   const [subMenuItem, setSubMenuItem] = useState<AppHeaderDropdownItem>();
 
@@ -38,6 +47,7 @@ export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = (
 
   const closeSubMenu = () => {
     setSubMenuItem(undefined);
+    getItemType(undefined);
   };
 
   const mapItemToElement = (
@@ -52,6 +62,7 @@ export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = (
       case 'dropdown':
       case 'profile-dropdown':
       case 'catalog-dropdown':
+      case 'experimental-resources-dropdown':
         return (
           <AppHeaderSubMenuTarget
             key={item.id}
@@ -97,15 +108,48 @@ export const AppHeaderMainMenuMobile: React.FC<AppHeaderMainMenuMobileProps> = (
       item={subMenuItem}
     />
   ) : (
-    <ContentContainer>
-      <AppHeaderListItem>
-        <MobileSearchBar onSearch={onSearch} />
-      </AppHeaderListItem>
-      {items.map((item) => (
-        <AppHeaderListItem key={item.id}>
-          {mapItemToElement(item, action)}
+    <Box>
+      <ContentContainer>
+        <AppHeaderListItem>
+          <MobileSearchBar onSearch={onSearch} />
         </AppHeaderListItem>
-      ))}
-    </ContentContainer>
+        {items.map((item) => (
+          <AppHeaderListItem key={item.id}>
+            {mapItemToElement(item, action)}
+          </AppHeaderListItem>
+        ))}
+      </ContentContainer>
+
+      {isAnon && (
+        <FlexBox
+          as="li"
+          alignItems="baseline"
+          justifyContent="center"
+          pt={16}
+          borderTop={1}
+          borderColor="navy-300"
+        >
+          <FillButton
+            data-testid={signUp.dataTestId}
+            href={signUp.href}
+            onClick={(event: React.MouseEvent) => action(event, signUp)}
+            key={signUp.id}
+            role="menuitem"
+          >
+            {signUp.text}
+          </FillButton>
+
+          <TextButton
+            key={login.id}
+            data-testid={login.dataTestId}
+            href={login.href}
+            onClick={(event: React.MouseEvent) => action(event, login)}
+            role="menuitem"
+          >
+            {login.text}
+          </TextButton>
+        </FlexBox>
+      )}
+    </Box>
   );
 };