+++ packages/gamut-labs/src/AppHeaderMobile/AppHeaderSubMenuMobile/index.tsx
@@ -1,4 +1,4 @@
-import { Anchor, ContentContainer, Text } from '@codecademy/gamut';
+import { Anchor, Text } from '@codecademy/gamut';
 import { ArrowChevronLeftIcon } from '@codecademy/gamut-icons';
 import { css } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
@@ -9,13 +9,27 @@ import { AppHeaderCatalogSection } from '../../AppHeader/AppHeaderElements/AppHe
 import { AppHeaderDropdownProps } from '../../AppHeader/AppHeaderElements/AppHeaderDropdown';
 import { AppHeaderLinkSections } from '../../AppHeader/AppHeaderElements/AppHeaderLinkSections';
 import { AppHeaderListItem } from '../../AppHeader/AppHeaderElements/AppHeaderListItem';
+import { AppHeaderResourceDropdownProps } from '../../AppHeader/AppHeaderElements/AppHeaderResourcesDropdown';
+import { AppHeaderResourcesSection } from '../../AppHeader/AppHeaderElements/AppHeaderResourcesSection';
+import {
+  AppHeaderCatalogDropdownItem,
+  AppHeaderClickHandler,
+  AppHeaderDropdownItem,
+  AppHeaderResourcesDropdownItem,
+} from '../../AppHeader/AppHeaderElements/types';
 
 export type AppHeaderSubMenuMobileProps = (
   | AppHeaderDropdownProps
   | AppHeaderCatalogDropdownProps
+  | AppHeaderResourceDropdownProps
 ) & {
   handleClose: () => void;
 };
+
+type AppHeaderSectionItem =
+  | AppHeaderDropdownItem
+  | AppHeaderCatalogDropdownItem
+  | AppHeaderResourcesDropdownItem;
 
 const StyledAnchor = styled(Anchor)(
   css({
@@ -27,6 +41,22 @@ const StyledAnchor = styled(Anchor)(
   })
 );
 
+const renderHeaderSection = (
+  item: AppHeaderSectionItem,
+  action: AppHeaderClickHandler
+) => {
+  switch (item.type) {
+    case 'catalog-dropdown':
+      return <AppHeaderCatalogSection action={action} item={item} />;
+    case 'experimental-resources-dropdown':
+      return <AppHeaderResourcesSection action={action} />;
+    default:
+      return (
+        <AppHeaderLinkSections action={action} item={item} showIcon mobile />
+      );
+  }
+};
+
 export const AppHeaderSubMenuMobile: React.FC<AppHeaderSubMenuMobileProps> = ({
   action,
   handleClose,
@@ -34,22 +64,26 @@ export const AppHeaderSubMenuMobile: React.FC<AppHeaderSubMenuMobileProps> = ({
 }) => {
   return (
     <AppHeaderListItem aria-labelledby={`${item.text} menu`}>
-      <ContentContainer>
-        <StyledAnchor onClick={handleClose} variant="interface" as="button">
-          <ArrowChevronLeftIcon size={12} aria-hidden />
-          <Text fontSize={16} ml={8}>
-            Full Menu
-          </Text>
-        </StyledAnchor>
-        <Text as="h1" fontSize={22} mb={16}>
-          {item.type === 'profile-dropdown' ? item.userDisplayName : item.text}
+      <StyledAnchor
+        onClick={handleClose}
+        variant="interface"
+        as="button"
+        ml={{ _: 16, xs: 32, sm: 64, md: 48 }}
+      >
+        <ArrowChevronLeftIcon size={12} aria-hidden />
+        <Text fontSize={16} pl={8}>
+          Menu
         </Text>
-      </ContentContainer>
-      {item.type === 'catalog-dropdown' ? (
-        <AppHeaderCatalogSection action={action} item={item} />
-      ) : (
-        <AppHeaderLinkSections action={action} item={item} showIcon mobile />
-      )}
+      </StyledAnchor>
+      <Text
+        as="h1"
+        fontSize={22}
+        mb={16}
+        ml={{ _: 16, xs: 32, sm: 64, md: 48 }}
+      >
+        {item.type === 'profile-dropdown' ? item.userDisplayName : item.text}
+      </Text>
+      {renderHeaderSection(item, action)}
     </AppHeaderListItem>
   );
 };