+++ packages/gamut-labs/src/GlobalHeader/GlobalHeaderVariants.tsx
@@ -20,6 +20,7 @@ import {
   pricingLink,
   proLogo,
   proProfile,
+  refreshedResourcesDropdown,
   resourcesDropdown,
   signUp,
   tryProForFree,
@@ -32,6 +33,9 @@ const catalogComponent = (user?: User) =>
   user?.useNewCatalogDropdown
     ? catalogDropdown(user?.hideCareerPaths)
     : courseCatalog;
+
+const resourcesComponent = (user?: User) =>
+  user?.useNewCatalogDropdown ? refreshedResourcesDropdown : resourcesDropdown;
 
 // Simplify pricing dropdown to a normal link for users in India
 const pricingComponent = (user?: User) =>
@@ -47,7 +51,7 @@ const anonHeaderItems = (
   const leftItems: AppHeaderItem[] = [
     logo,
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     ...(hidePricing ? [] : [pricingComponent(user)]),
     businessSolutions,
@@ -92,12 +96,10 @@ const anonMobileHeaderItems = (
 
   const mainMenuItems: AppHeaderItem[] = [
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     ...(hidePricing ? [] : [pricingComponent(user)]),
     businessSolutions,
-    signUp,
-    login,
   ];
 
   return {
@@ -181,7 +183,7 @@ export const freeHeaderItems = (
     logo,
     myHome,
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     ...(hidePricing ? [] : [pricingComponent(user)]),
     businessSolutions,
@@ -218,7 +220,7 @@ export const freeMobileHeaderItems = (
   const mainMenuItems: AppHeaderItem[] = [
     myHome,
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     ...(hidePricing ? [] : [pricingComponent(user)]),
     businessSolutions,
@@ -248,7 +250,7 @@ export const proHeaderItems = (
     user.hasNewSkuSubscription ? logo : proLogo,
     myHome,
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     businessSolutions,
   ];
@@ -284,7 +286,7 @@ export const proMobileHeaderItems = (
   const mainMenuItems: AppHeaderItem[] = [
     myHome,
     catalogComponent(user),
-    resourcesDropdown,
+    resourcesComponent(user),
     communityDropdown,
     businessSolutions,
     proProfile(user),