+++ packages/gamut-labs/src/GlobalHeader/index.tsx
@@ -155,6 +155,7 @@ export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
           : {
               notifications: props.notifications,
             })}
+        isAnon={props.type === 'anon'}
       />
       <AppHeaderMobile
         action={combinedAction}
@@ -166,6 +167,7 @@ export const GlobalHeader: React.FC<GlobalHeaderProps> = (props) => {
             })}
         onSearch={props.search.onSearch}
         redirectParam={props.type === 'anon' ? props.redirectParam : undefined}
+        isAnon={props.type === 'anon'}
       />
       {props.children}
     </Box>