+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [58.1.10](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.9...@codecademy/styleguide@58.1.10) (2022-07-08)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [58.1.9](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.8...@codecademy/styleguide@58.1.9) (2022-07-06)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [58.1.8](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.7...@codecademy/styleguide@58.1.8) (2022-07-06)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [58.1.7](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.6...@codecademy/styleguide@58.1.7) (2022-06-30)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [58.1.6](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.5...@codecademy/styleguide@58.1.6) (2022-06-28)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
 ### [58.1.5](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@58.1.4...@codecademy/styleguide@58.1.5) (2022-06-27)
 
 **Note:** Version bump only for package @codecademy/styleguide